fix: label fix and refactoring
