+++ library/lib/edges/GenericEdge.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useRef, useState, useEffect } from "react"
+import { useRef, useState, useEffect } from "react"
 import { useReactFlow, type Node } from "@xyflow/react"
 import { ExtendedEdgeProps } from "./EdgeProps"
 import { CustomEdgeToolbar } from "@/components"
@@ -10,7 +10,7 @@ import { DiagramEdgeType } from "."
 import { Assessment } from "@/typings"
 
 export interface BaseEdgeProps extends ExtendedEdgeProps {
-  diagramType?: "class" | "usecase" | "activity"
+  diagramType?: "class" | "usecase" | "activity" | "component" | "deployment"
 }
 export const useEdgeState = (initialPoints?: IPoint[]) => {
   const [customPoints, setCustomPoints] = useState<IPoint[]>([])
@@ -29,93 +29,6 @@ export const useEdgeState = (initialPoints?: IPoint[]) => {
     setCustomPoints,
     tempReconnectPoints,
     setTempReconnectPoints,
-  }
-}
-export const useEdgePath = (
-  sourceX: number,
-  sourceY: number,
-  targetX: number,
-  targetY: number,
-  pathRef: React.RefObject<SVGPathElement>
-) => {
-  const [pathMiddlePosition, setPathMiddlePosition] = useState<IPoint>(() => ({
-    x: (sourceX + targetX) / 2,
-    y: (sourceY + targetY) / 2,
-  }))
-
-  const [isMiddlePathHorizontal, setIsMiddlePathHorizontal] = useState<boolean>(
-    () => {
-      const dx = Math.abs(targetX - sourceX)
-      const dy = Math.abs(targetY - sourceY)
-      return dx > dy
-    }
-  )
-
-  useEffect(() => {
-    const middleX = (sourceX + targetX) / 2
-    const middleY = (sourceY + targetY) / 2
-    setPathMiddlePosition({ x: middleX, y: middleY })
-
-    const dx = Math.abs(targetX - sourceX)
-    const dy = Math.abs(targetY - sourceY)
-    setIsMiddlePathHorizontal(dx > dy)
-  }, [sourceX, sourceY, targetX, targetY])
-
-  const updateMiddlePosition = useCallback(
-    (isDirectPath = false) => {
-      if (isDirectPath) {
-        const middleX = (sourceX + targetX) / 2
-        const middleY = (sourceY + targetY) / 2
-        setPathMiddlePosition({ x: middleX, y: middleY })
-
-        const dx = Math.abs(targetX - sourceX)
-        const dy = Math.abs(targetY - sourceY)
-        setIsMiddlePathHorizontal(dx > dy)
-      } else if (pathRef.current) {
-        try {
-          const totalLength = pathRef.current.getTotalLength()
-          if (totalLength === 0 || !isFinite(totalLength)) {
-            const middleX = (sourceX + targetX) / 2
-            const middleY = (sourceY + targetY) / 2
-            setPathMiddlePosition({ x: middleX, y: middleY })
-
-            const dx = Math.abs(targetX - sourceX)
-            const dy = Math.abs(targetY - sourceY)
-            setIsMiddlePathHorizontal(dx > dy)
-            return
-          }
-
-          const halfLength = totalLength / 2
-          const middlePoint = pathRef.current.getPointAtLength(halfLength)
-          const pointOnCloseToMiddle = pathRef.current.getPointAtLength(
-            Math.min(halfLength + 2, totalLength)
-          )
-
-          const isHorizontal =
-            Math.abs(pointOnCloseToMiddle.x - middlePoint.x) >
-            Math.abs(pointOnCloseToMiddle.y - middlePoint.y)
-
-          setIsMiddlePathHorizontal(isHorizontal)
-          setPathMiddlePosition({ x: middlePoint.x, y: middlePoint.y })
-        } catch (error) {
-          console.warn("Path calculation failed, using fallback:", error)
-          const middleX = (sourceX + targetX) / 2
-          const middleY = (sourceY + targetY) / 2
-          setPathMiddlePosition({ x: middleX, y: middleY })
-
-          const dx = Math.abs(targetX - sourceX)
-          const dy = Math.abs(targetY - sourceY)
-          setIsMiddlePathHorizontal(dx > dy)
-        }
-      }
-    },
-    [sourceX, sourceY, targetX, targetY, pathRef]
-  )
-
-  return {
-    updateMiddlePosition,
-    pathMiddlePosition,
-    isMiddlePathHorizontal,
   }
 }
