+++ library/lib/edges/pathTypes/StraightPathEdge.tsx
@@ -1,7 +1,7 @@
-import { useMemo, useRef } from "react"
+import { useEffect, useMemo, useRef, useState } from "react"
 import React from "react"
 import { BaseEdge } from "@xyflow/react"
-import { BaseEdgeProps, useEdgePath, CommonEdgeElements } from "../GenericEdge"
+import { BaseEdgeProps, CommonEdgeElements } from "../GenericEdge"
 import {
   calculateOverlayPath,
   calculateStraightPath,
@@ -45,19 +45,23 @@ export const StraightPathEdge = ({
     }))
   )
 
+  const [pathMiddlePosition, setPathMiddlePosition] = useState<IPoint>(() => ({
+    x: (sourceX + targetX) / 2,
+    y: (sourceY + targetY) / 2,
+  }))
+  const [isMiddlePathHorizontal, setIsMiddlePathHorizontal] = useState<boolean>(
+    () => {
+      const dx = Math.abs(targetX - sourceX)
+      const dy = Math.abs(targetY - sourceY)
+      return dx > dy
+    }
+  )
+
   const setPopOverElementId = usePopoverStore(
     useShallow((state) => state.setPopOverElementId)
   )
 
   const { handleDelete } = useToolbar({ id })
-
-  const { pathMiddlePosition, isMiddlePathHorizontal } = useEdgePath(
-    sourceX,
-    sourceY,
-    targetX,
-    targetY,
-    pathRef
-  )
 
   const { markerEnd, strokeDashArray } = getEdgeMarkerStyles(type)
 
@@ -69,14 +73,59 @@ export const StraightPathEdge = ({
     return calculateOverlayPath(sourceX, sourceY, targetX, targetY, type)
   }, [sourceX, sourceY, targetX, targetY, type])
 
+  useEffect(() => {
+    if (pathRef.current) {
+      try {
+        const totalLength = pathRef.current.getTotalLength()
+        if (totalLength === 0 || !isFinite(totalLength)) {
+          const middleX = (sourceX + targetX) / 2
+          const middleY = (sourceY + targetY) / 2
+          setPathMiddlePosition({ x: middleX, y: middleY })
+
+          const dx = Math.abs(targetX - sourceX)
+          const dy = Math.abs(targetY - sourceY)
+          setIsMiddlePathHorizontal(dx > dy)
+          return
+        }
+
+        const halfLength = totalLength / 2
+        const middlePoint = pathRef.current.getPointAtLength(halfLength)
+        const pointOnCloseToMiddle = pathRef.current.getPointAtLength(
+          Math.min(halfLength + 2, totalLength)
+        )
+        const isHorizontal =
+          Math.abs(pointOnCloseToMiddle.x - middlePoint.x) >
+          Math.abs(pointOnCloseToMiddle.y - middlePoint.y)
+
+        setIsMiddlePathHorizontal(isHorizontal)
+        setPathMiddlePosition({ x: middlePoint.x, y: middlePoint.y })
+      } catch (error) {
+        console.warn("Path calculation failed, using fallback:", error)
+        // Fallback to simple midpoint calculation
+        const middleX = (sourceX + targetX) / 2
+        const middleY = (sourceY + targetY) / 2
+        setPathMiddlePosition({ x: middleX, y: middleY })
+
+        const dx = Math.abs(targetX - sourceX)
+        const dy = Math.abs(targetY - sourceY)
+        setIsMiddlePathHorizontal(dx > dy)
+      }
+    }
+  }, [currentPath, sourceX, sourceY, targetX, targetY])
+
+  // Update position immediately when coordinates change (for smooth transitions)
+  useEffect(() => {
+    const middleX = (sourceX + targetX) / 2
+    const middleY = (sourceY + targetY) / 2
+    setPathMiddlePosition({ x: middleX, y: middleY })
+
+    const dx = Math.abs(targetX - sourceX)
+    const dy = Math.abs(targetY - sourceY)
+    setIsMiddlePathHorizontal(dx > dy)
+  }, [sourceX, sourceY, targetX, targetY])
+
   const sourcePoint = { x: sourceX, y: sourceY }
   const targetPoint = { x: targetX, y: targetY }
-  const edgeData: StraightPathEdgeData = {
-    pathMiddlePosition,
-    isMiddlePathHorizontal,
-    sourcePoint,
-    targetPoint,
-  }
 
   return (
     <>
@@ -103,43 +152,29 @@ export const StraightPathEdge = ({
         />
       </g>
 
-      {typeof children === "function" ? children(edgeData) : children}
+      {pathMiddlePosition && (
+        <>
+          {typeof children === "function"
+            ? children({
+                pathMiddlePosition,
+                isMiddlePathHorizontal,
+                sourcePoint,
+                targetPoint,
+              })
+            : children}
 
-      <CommonEdgeElements
-        id={id}
-        pathMiddlePosition={pathMiddlePosition}
-        isDiagramModifiable={isDiagramModifiable}
-        assessments={assessments}
-        anchorRef={anchorRef}
-        handleDelete={handleDelete}
-        setPopOverElementId={setPopOverElementId}
-        type={type}
-      />
+          <CommonEdgeElements
+            id={id}
+            pathMiddlePosition={pathMiddlePosition}
+            isDiagramModifiable={isDiagramModifiable}
+            assessments={assessments}
+            anchorRef={anchorRef}
+            handleDelete={handleDelete}
+            setPopOverElementId={setPopOverElementId}
+            type={type}
+          />
+        </>
+      )}
     </>
   )
-}
-
-export const useStraightPathData = (
-  straightPathEdgeProps: StraightPathEdgeProps
-) => {
-  const { pathMiddlePosition, isMiddlePathHorizontal } = useEdgePath(
-    straightPathEdgeProps.sourceX,
-    straightPathEdgeProps.sourceY,
-    straightPathEdgeProps.targetX,
-    straightPathEdgeProps.targetY,
-    { current: null }
-  )
-
-  return {
-    pathMiddlePosition,
-    isMiddlePathHorizontal,
-    sourcePoint: {
-      x: straightPathEdgeProps.sourceX,
-      y: straightPathEdgeProps.sourceY,
-    },
-    targetPoint: {
-      x: straightPathEdgeProps.targetX,
-      y: straightPathEdgeProps.targetY,
-    },
-  }
 }
